pyproject.toml
[build-system]
requires = [
    "pip>=22.0.4",
    "pytest-runner>=6.0.0",
    "setuptools>=61.2.0",
    "setuptools_scm[toml]>=6.4.2",
    "wheel>=0.37.1",
]
build-backend = "setuptools.build_meta"

[project]
authors = [
    { name = "Jason Lubken", email = "jlubken@users.noreply.github.com" },
    { name = "Michael Becker", email = "mdbecker@users.noreply.github.com" },
    { name = "Corey Chivers", email = "cjbayesian@user.noreply.github.com" },
    { name = "Graydon Neill", email = "grayeye@user.noreply.github.com"},
    { name = "Penn Signals", email = "pennsignals@pennmedicine.upenn.edu" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Unix",
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Utilities",
]
dependencies = [
    "blosc>=1.10.6",
    "cfgenvy>=1.4.2",
    "numpy>=1.15.4",
    "pandas>=0.23.4",
    "requests>=2.26.0",
]
description = "An opinionated library to help deploy data science projects"
dynamic = ["version"]
license = { file = "license.txt" }
name = "dsdk"
requires-python = ">=3.7"
readme = "readme.md"

[project.optional-dependencies]
dev = [
    "astroid",
    "black",
    "coverage[toml]",
    "cython>=0.29.21",
    "flake8",
    "flake8-bugbear",
    "flake8-commas",
    "flake8-comprehensions",
    "flake8-docstrings",
    # "flake8-logging-format",
    "flake8-mutable",
    "flake8-sorted-keys",
    "isort",
    "mypy",
    "pep8-naming",
    "pre-commit",
    "psycopg2-binary>=2.8.6",
    "pylint",
    "pymssql>=2.2.3",
    "pytest",
    "pytest-cov",
    "types-pkg-resources",
    "types-pymssql",
    "types-python-dateutil",
    "types-pymssql",
    "types-pyyaml",
    "types-requests",
    "vcrpy",
]
pymssql = [
    "cython>=0.29.21",
    "pymssql>=2.2.3",
]
psycopg2 = [
    "psycopg2-binary>=2.8.6",
]

[project.urls]
Homepage = "https://github.com/pennsignals/dsdk"
Source = "https://github.com/pennsignals/dsdk"
Tracker = "https://github.com/pennsignals/dsdk/issues"

[tool.black]
line-length = 79
target-version = ["py37","py38","py39","py310"]

[tool.coverage.report]
exclude_lines = [
    "raise NotImplementedError()",
    "pragma: nocover",
    "if __name__ == .__main__.:",
]
precision = 2
show_missing = true

[tool.coverage.run]
branch = true
parallel = true

[tool.distutils.bdist_wheel]
python-tag = "py37.py38.py39.py310"

[tool.isort]
profile = "black"
include_trailing_comma = true
line_length = 79
multi_line_output = 3
known_first_party = ["dsdk"]
default_section = "THIRDPARTY"

[tool.mypy]
follow_imports = "normal"
ignore_missing_imports = false

[tool.pylint.basic]
good-names = '''a,b,c,d,df,do_POST,e,i,id,logger,n,on,ok,tz'''

[tool.pylint.message_control]
disable = '''duplicate-code'''
extension-pkg-allow-list = [
  "pymssql._mssql",
]

[tool.pylint.miscellaneous]
notes = '''FIXME,XXX'''

[tool.pylint.similarities]
min-similarity-lines = 4
ignore-comments = "yes"
ignore-docstrings = "yes"
ignore-imports = "yes"

[tool.pytest.ini_options]
addopts = "-ra --cov=dsdk --cov-report=term-missing --strict-markers --ignore=.eggs --tb=short"
testpaths = ["test"]
norecursedirs = ".env .git .venv build dist"
python_files = "test.py tests.py test_*.py *_test.py"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
local_scheme = "dirty-tag"
